refactor(server): recreate userModule and move related functions back into it
