+++ packages/core/bin/irving
@@ -1,5 +1,9 @@
 #!/usr/bin/env node
 /* eslint-disable global-require, max-len */
+
+// Initialize env vars as soon as possible for the CLI.
+require('../config/env')();
+
 ['react', 'react-dom'].forEach((dependency) => {
   try {
     // When 'npm link' is used it checks the clone location. Not the project.
